views dir removed
