+++ src/components/Common/Header/index.ts
@@ -1,2 +1,3 @@
 import Header from './Header';
-export { Header };
+
+export default Header;
