+++ src/components/Dashboard/Tiles/Date/index.ts
@@ -1,0 +1,3 @@
+import DateTile from './DateTile';
+
+export default DateTile;
