+++ src/components/Dashboard/Tiles/HalfHourCount/index.ts
@@ -1,0 +1,3 @@
+import HalfHourCount from './HalfHourCount';
+
+export default HalfHourCount;
