+++ src/components/Dashboard/Tiles/HalfHourCount/styled.tsx
