+++ src/components/Dashboard/Tiles/MaskChart/index.ts
@@ -1,0 +1,3 @@
+import MaskPercTile from './MaskPercTile';
+
+export default MaskPercTile;
