+++ src/components/Dashboard/Tiles/RecentUsers/index.ts
@@ -1,0 +1,3 @@
+import RecentUsers from './RecentUsers';
+
+export default RecentUsers;
