+++ src/components/Dashboard/Tiles/TotalUsers/styled.tsx
@@ -1,0 +1,25 @@
+import styled from 'styled-components';
+
+export const TileContainer = styled.div`
+  height: 130px;
+  width: 10rem;
+  background-color: #c73749;
+  box-shadow: 2px 2px gainsboro;
+  padding: 15px;
+  border-radius: 10px;
+  text-align: left;
+`;
+
+export const CountTitle = styled.p`
+  font-size: 25px;
+  color: #ffffff;
+  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen,
+    Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
+  margin: 0;
+`;
+
+export const NumberDisplay = styled.div`
+  font-size: 35px;
+  margin: 20px;
+  color: #ffffff;
+`;
