+++ src/components/Dashboard/Tiles/UsersGraph/index.ts
@@ -1,0 +1,3 @@
+import UsersGraph from './UsersGraph';
+
+export default UsersGraph;
