+++ src/components/Dashboard/index.ts
@@ -1,14 +1,3 @@
-import RecentUsers from './RecentUsers/RecentUsers';
-import TotalUsers from './TotalUsers/TotalUsers';
-import UsersGraph from './UsersGraph/UsersGraph';
-import HalfHourCount from './HalfHourCount/HalfHourCount';
-import DateTile from './Date/DateTile';
-import MaskPercTile from './MaskChart/MaskPercTile';
-export {
-  RecentUsers,
-  TotalUsers,
-  HalfHourCount,
-  UsersGraph,
-  DateTile,
-  MaskPercTile,
-};
+import Dashboard from './Dashboard';
+
+export default Dashboard;
