+++ src/components/Livestream/Livestream.tsx
@@ -1,15 +1,5 @@
 import React from 'react';
-import styled from 'styled-components';
-
-const Display = styled.div`
-  padding: 10px;
-  display: flex;
-  justify-content: start;
-  align-items: flex-start;
-  flex-direction: column;
-  max-height: 90vh;
-  gap: 10px;
-`;
+import { Display } from './styled';
 
 const Livestream = () => {
   return (
@@ -25,4 +15,4 @@ const Livestream = () => {
   );
 };
 
-export { Livestream };
+export default Livestream;
