+++ src/components/Livestream/index.ts
@@ -1,0 +1,3 @@
+import Livestream from './Livestream';
+
+export default Livestream;
