+++ src/components/Settings/Settings.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+import { Display } from './styled';
+
+const Settings = () => {
+  return <Display>Settings</Display>;
+};
+
+export default Settings;
