+++ src/components/Sidebar/index.ts
@@ -1,2 +1,3 @@
 import Sidebar from './Sidebar';
-export { Sidebar };
+
+export default Sidebar;
