+++ src/components/Users/Users.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+import { Display } from './styled';
+
+const Users = () => {
+  return <Display>Users</Display>;
+};
+
+export default Users;
