+++ src/components/Users/styled.tsx
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+export const Display = styled.div`
+  padding: 10px;
+  display: flex;
+  justify-content: start;
+  align-items: flex-start;
+  flex-direction: column;
+  max-height: 90vh;
+  gap: 10px;
+`;
