--- src/components/index.ts
@@ -1,19 +1,0 @@
-import { Sidebar } from './Sidebar';
-import {
-  RecentUsers,
-  TotalUsers,
-  UsersGraph,
-  HalfHourCount,
-  DateTile,
-  MaskPercTile,
-} from './Dashboard';
-
-export {
-  Sidebar,
-  RecentUsers,
-  TotalUsers,
-  UsersGraph,
-  HalfHourCount,
-  DateTile,
-  MaskPercTile,
-};
