--- src/views/index.ts
@@ -1,6 +1,0 @@
-import { Dashboard } from '../components/Dashboard/Dashboard';
-import { Livestream } from './Livestream';
-import { Settings } from './Settings';
-import { Users } from './Users';
-
-export { Dashboard, Livestream, Settings, Users };
